* commit '83e0818ca37273b99b5f43ce0dcddb2b316b22b9': (92 commits)
  Updated doxygen output
  Squashed 'epc9143-avmc.X/sources/power_control/devices/' content from commit 7c88e270
  Renamed power converter directory from pwr_control to power_control
  Updated subtree repository management scripts
  fixing typo in meta file
  Fixed missing paranthesis in meta file
  Updated firmware version tag
  Fixed typo in meta file causing server build to fail
  Updated changelog fixing typo
  Updated meta file keywords
  Updated README file adding safety requirements paragraph
  Updated firmware documentation of version 2.2
  Updated configuration for code optimization level #1 and enabled debugger/programmer power for offline programming
  Added complete output current calcualtion to power supply state machine
  Reordered ADC triggers of VIN, VOUT and ISNS1 + ISNS2
  Added missing current feedback settings and new switch-node settings to power supply configuration
  Reordering documentation index
  Added return value to high priority task function call to keep function call active when code optimization is used
  Updated gitignore file to include documentation directory
  Updated documentation output after reordering index of hardware decription layer and moving PRAL drivers to Library Drivers
  ...

+++ .mchp_private/doxygen_src/build-output.bat
@@ -48,6 +48,7 @@ echo skipped
 goto skip_delete_output
 :delete_output
 del /F /Q *.*
+if exist ".\search" del /F /Q search\*.*
 echo done
 :skip_delete_output
 echo. 
@@ -58,7 +59,7 @@ set yesno="n"
 echo skipped
 goto skip_delete_images
 :delete_images
-del /F /Q images\*.*
+if exist ".\images" del /F /Q images\*.*
 echo done
 :skip_delete_images
 echo. 
@@ -69,13 +70,18 @@ set yesno="n"
 echo skipped
 :generate_output
 echo. 
+echo ##############  R U N   D O X Y G E N  ##############
 echo. 
 echo Calling Doxygen Generator...
 cd ..\.mchp_private\doxygen_src
 doxygen doxyfile
 echo. 
-echo Copying Image Folder:
+echo ##############  D O X Y G E N   E X E C U T I O N   E N D  ##############
+echo. 
+echo Merging Template Image Folders:
 Xcopy /e /i /c /r /y ..\..\images ..\..\docs\images
+if exist ".\markdown-files\images" Xcopy /e /i /c /r /y .\markdown-files\images ..\..\docs\images
+if exist ".\images" Xcopy /e /i /c /r /y .\images ..\..\docs\images
 echo. 
 echo ##############  D O N E  ##############
 echo.
